# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Budget API",
        "description": "API endpoints for managing budgets",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/prices": {
            "get": {
                "tags": [
                    "Prices"
                ],
                "summary": "Get all prices",
                "operationId": "f43226185d5e1c63f1dc1240b9b9f988",
                "responses": {
                    "200": {
                        "description": "List of prices retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Price"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Prices"
                ],
                "summary": "Create a new price",
                "operationId": "aee210ae6657c2d401f25617b65bbc1a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PriceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Price created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Price"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/prices/{id}": {
            "get": {
                "tags": [
                    "Prices"
                ],
                "summary": "Get price by ID",
                "operationId": "9f2c07a5e217eceb2e162042af79aa39",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Price ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Price found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Price"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Price not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Prices"
                ],
                "summary": "Update an existing price",
                "operationId": "ca021243deee9acc438bcadb6177af03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PriceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Price updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Price"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Prices"
                ],
                "summary": "Delete a price",
                "operationId": "babf67601c6984784fb23172de05732f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Price deleted successfully"
                    },
                    "404": {
                        "description": "Price not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Price": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "price": {
                        "type": "number",
                        "format": "float"
                    },
                    "date": {
                        "type": "string",
                        "format": "date"
                    },
                    "material_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "PriceRequest": {
                "required": [
                    "price",
                    "date",
                    "material_id"
                ],
                "properties": {
                    "price": {
                        "type": "number",
                        "format": "float"
                    },
                    "date": {
                        "type": "string",
                        "format": "date"
                    },
                    "material_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}